requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-groq
langchain-tavily
arxiv
wikipedia
